endermen/requirements.txt
python-ffmpeg
pandas
numpy
matplotlib
seaborn